Fixed an issue with internal ServiceRequest from SvnOpenFile: on-demand authentication was only enabled if the user was already logged in. This meant pages using SvnOpenFile could only respond to access control changes but did not support clients that had yet to authenticate.
Also added an automatic HEAD request to the target for all SvnOpenFile initializations. This is because a lot of pages only initialize SvnOpenFile, then send it to the template, resulting in errors instead of authentication because page body has already started when the first read operation is executed. The API is more intuitive if the caller need not care about this. Unit tests must disable this automatic functionality.

git-svn-id: https://optime.repos.se/svn/code/repos/trunk@4719 94f62164-8fde-0310-83d7-e1105c8c1596