Experimented with image display based on solr index. Using Galleriffic 2.0.
Created a new js bundle, libs.js, for the shared scripts that Repos and plugins can expect to be loaded. This would make it easier to include the base package without including all plugins (maybe Repos API should be there too). On the other hand it does not force plugin pages to specify which libs they use.

git-svn-id: https://optime.repos.se/svn/code/repos/trunk@5388 94f62164-8fde-0310-83d7-e1105c8c1596